requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
aiofiles==23.2.1
pydantic==2.4.2
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
